Features/general

+++ src/main/java/com/qa/ims/controller/CustomerController.java
@@ -47,9 +47,8 @@ public class CustomerController implements CrudController<Customer> {
 		String firstName = utils.getString();
 		LOGGER.info("Please enter a surname: ");
 		String surname = utils.getString();
-		// enter address once everything is working
 		Customer customer = customerDAO.create(new Customer(firstName, surname));
-		LOGGER.info("Customer created: ");
+		LOGGER.info("Customer created.");
 		return customer;
 	}
 
@@ -58,8 +57,11 @@ public class CustomerController implements CrudController<Customer> {
 	 */
 	@Override
 	public Customer update() {
-		LOGGER.info("Please enter the id of the customer you would like to update: ");
+		LOGGER.info("Please enter the ID of the customer you would like to update: ");
 		Long id = utils.getLong();
+
+		if (!checkID(id)) return null;
+
 		LOGGER.info("Please enter a first name: ");
 		String firstName = utils.getString();
 		LOGGER.info("Please enter a surname: ");
@@ -67,18 +69,34 @@ public class CustomerController implements CrudController<Customer> {
 		Customer customer = customerDAO.update(new Customer(id, firstName, surname));
 		LOGGER.info("Customer successfully updated.");
 		return customer;
+
 	}
 
 	/**
-	 * Deletes an existing customer by the id of the customer
+	 * Deletes an existing customer using the id of the customer
 	 * 
 	 * @return
 	 */
 	@Override
 	public int delete() {
-		LOGGER.info("Please enter the id of the customer you would like to delete");
+		LOGGER.info("Please enter the ID of the customer you would like to delete: ");
 		Long id = utils.getLong();
+
+		if (!checkID(id)) return 0;
+
 		return customerDAO.delete(id);
+	}
+
+	private boolean checkID(Long id) {
+		List<Customer> customers = customerDAO.readAll();
+		int count = 0;
+		for (Customer customer : customers){
+			if (customer.getID() == id) count++;
+		}
+		if (count == 0) {
+			LOGGER.info("Customer does not exist");
+			return false;
+		} else return true;
 	}
 
 }
